# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medical Records API",
        "description": "API quan ly ho so y te",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Danh sách người dùng",
                "operationId": "fbe4abc6115f6a931262eefde634d1f9",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "schema": {
                            "description": "Tìm kiếm theo tên đăng nhập",
                            "type": "string"
                        }
                    },
                    {
                        "name": "staff_id",
                        "in": "query",
                        "schema": {
                            "description": "ID nhân viên liên kết",
                            "type": "string"
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "schema": {
                            "description": "ID bệnh nhân liên kết",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách người dùng"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Tạo người dùng mới",
                "operationId": "243db3d4e5f82ed07ca14dc2bf82410c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "doctor01"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "doctor01@hospital.com"
                                    },
                                    "password_hash": {
                                        "description": "Password đã được hash bằng bcrypt",
                                        "type": "string",
                                        "example": "$2y$10$..."
                                    },
                                    "role_names": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "doctor",
                                            "user"
                                        ]
                                    },
                                    "account_type": {
                                        "type": "string",
                                        "enum": [
                                            "staff",
                                            "patient"
                                        ],
                                        "example": "staff"
                                    },
                                    "linked_staff_id": {
                                        "type": "string",
                                        "example": "staff_123"
                                    },
                                    "linked_patient_id": {
                                        "type": "string",
                                        "example": "patient_456"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "422": {
                        "description": "Lỗi validation"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Chi tiết người dùng",
                "operationId": "05dabd18926420b0d890e4fc40b53825",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin người dùng"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Cập nhật người dùng",
                "operationId": "f1b6ca0c6427f48f79ba9e3abc6c4a42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "_rev": {
                                        "type": "string",
                                        "example": "1-abc123"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "doctor02"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "doctor02@hospital.com"
                                    },
                                    "role_names": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Xóa người dùng",
                "operationId": "d2cba55c56618a8c24287a7fe2b89a60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rev",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "400": {
                        "description": "Thiếu rev parameter"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        }
    },
...
}